Renamed map javascript files.
