+++ app/javascript/source/show_donation_map.js
